Todo with userId

+++ apps/api/src/api/todos/todos.controller.ts
@@ -9,9 +9,11 @@ import {
   Patch,
   Post,
   Query,
+  Req,
   UseGuards,
 } from '@nestjs/common';
 import { ApiTags } from '@nestjs/swagger';
+import { RequestWithAuth } from 'src/dto/base.dto';
 import { ResponseType } from 'src/dto/enums.dto';
 import { TaskQueryDto, TaskStoreDto, TaskUpdateDto } from 'src/dto/tasks.dto';
 import { TodosRepository } from 'src/repositories/todos/todos.repository';
@@ -29,8 +31,9 @@ export class TodosController {
 
   @Get()
   @HttpCode(HttpStatus.OK)
-  async index(@Query() queries: TaskQueryDto) {
-    const res = await this.repo.getDocs({ ...queries });
+  async index(@Query() queries: TaskQueryDto, @Req() req: RequestWithAuth) {
+    const { id: userId } = req.auth;
+    const res = await this.repo.getDocs({ ...queries, userId });
     if (!res.ok && res.type === ResponseType.SERVER_ERROR) {
       return this.helper.serverException(res.data as Error);
     }
@@ -42,8 +45,9 @@ export class TodosController {
 
   @Post()
   @HttpCode(HttpStatus.CREATED)
-  async store(@Body() body: TaskStoreDto) {
-    const res = await this.repo.createDoc(body);
+  async store(@Body() body: TaskStoreDto, @Req() req: RequestWithAuth) {
+    const { id: userId } = req.auth;
+    const res = await this.repo.createDoc({ ...body, userId });
     if (!res.ok && res.type === ResponseType.SERVER_ERROR) {
       return this.helper.serverException(res.data as Error);
     }
@@ -55,8 +59,13 @@ export class TodosController {
 
   @Get(':id')
   @HttpCode(HttpStatus.OK)
-  async show(@Param('id') id: string, @Query() queries: { id_key?: string }) {
-    const res = await this.repo.getDoc(id, { ...queries });
+  async show(
+    @Param('id') id: string,
+    @Query() queries: { id_key?: string },
+    @Req() req: RequestWithAuth,
+  ) {
+    const { id: userId } = req.auth;
+    const res = await this.repo.getDoc(id, { ...queries, userId });
     if (!res.ok && res.type === ResponseType.SERVER_ERROR) {
       return this.helper.serverException(res.data as Error);
     }
@@ -68,8 +77,13 @@ export class TodosController {
 
   @Patch(':id')
   @HttpCode(HttpStatus.OK)
-  async update(@Param('id') id: string, @Body() body: TaskUpdateDto) {
-    const res = await this.repo.updateDoc(id, body, {});
+  async update(
+    @Param('id') id: string,
+    @Body() body: TaskUpdateDto,
+    @Req() req: RequestWithAuth,
+  ) {
+    const { id: userId } = req.auth;
+    const res = await this.repo.updateDoc(id, body, { userId });
     if (!res.ok && res.type === ResponseType.SERVER_ERROR) {
       return this.helper.serverException(res.data as Error);
     }
@@ -81,8 +95,9 @@ export class TodosController {
 
   @Delete(':id')
   @HttpCode(HttpStatus.OK)
-  async destroy(@Param('id') id: string) {
-    const res = await this.repo.deleteDoc(id, {});
+  async destroy(@Param('id') id: string, @Req() req: RequestWithAuth) {
+    const { id: userId } = req.auth;
+    const res = await this.repo.deleteDoc(id, { userId });
     if (!res.ok && res.type === ResponseType.SERVER_ERROR) {
       return this.helper.serverException(res.data as Error);
     }
